Linting, Gatsby config, and update components
Adding eslinting to catch some stupid mistakes I make. Realized what I
was doing wrong with the imported components and so fixed it in the
repo. But to use styled components in a dependency and here in Gatsby, I
had to make a slight change to the config.

I also upgraded all the things while I was here.